refactor(225): update views architecture for future unit tests
