+++ src/features/student/views/StudentAboutView/StudentAboutView.vue
