+++ src/features/student/views/StudentEducationActivitiesView/StudentEducationActivitiesView.vue
