+++ src/features/student/views/StudentEducationSkillsView/StudentEducationSkillsView.vue
