+++ src/features/student/views/StudentEventsView/StudentEventsView.vue
