+++ src/features/student/views/StudentMailboxView/StudentMailboxView.vue
